SpreadsheetEngineHateosHandlerSpreadsheetDelta.selection pulled SpreadsheetEngineHateosHandler
